layout/style/test/chrome/test_bug1160724.xhtml

<?xml version="1.0"?>
<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<?xml-stylesheet type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"?>

<?xml-stylesheet href="data:text/css,:root{--test:9px}" type="text/css"?>

<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1160724
-->
<window title="Mozilla Bug 1160724" onload="test()"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=1160724"
     target="_blank">Mozilla Bug 1160724</a>
  </body>

  <script type="application/javascript">
  <![CDATA[
  var errorLogged = false;
  const serv = Cc["@mozilla.org/consoleservice;1"]
                 .getService(Ci.nsIConsoleService);
  var listener = {
    QueryInterface(iid) {
      if (!iid.equals(Ci.nsISupports) &&
          !iid.equals(Ci.nsIConsoleListener)) {
        throw Components.Exception("", Cr.NS_NOINTERFACE);
      }
      return this;
    },

    observe(msg) {
      if (msg.toString().includes("transform")) {
        errorLogged = true;
      }
    }
  };
  serv.registerListener(listener);
  ]]>
  </script>

  <vbox id="w" style="-moz-binding: url(#binding)">
    <vbox id="v" style="display: none; transform: translateY(var(--test));" />
  </vbox>

  <bindings xmlns="http://www.mozilla.org/xbl">
    <binding id="binding">
      <implementation>
        <constructor>this.firstChild</constructor>
      </implementation>
    </binding>
  </bindings>

  <!-- test code goes here -->
  <script type="application/javascript">
  <![CDATA[
  /** Test for Bug 1160724 **/
  SimpleTest.waitForExplicitFinish();

  function test() {
    var v = document.getElementById("v");
    is(getComputedStyle(v, "").transform, "matrix(1, 0, 0, 1, 0, 9)");

    // nsIConsoleListeners are notified by a runnable.
    setTimeout(() => {
      ok(!errorLogged, "Should be no errors");
      serv.unregisterListener(listener);
      SimpleTest.finish();
    })
  }
  ]]>
  </script>
</window>
